modify

+++ HARDWARE/RS232_protocol/slave_uart_control_interface.h
@@ -18,7 +18,7 @@
 #include "check_rs232.h"
 #include "export_liquid.h"
 
-uint8_t Rest_C55_C52_Position(void);
+uint32_t Rest_C55_C52_Position(void);
